Debugging Functionality for Training
Current Bug:
- Somewhere, the output array is getting NaN values, needs to be fixed

+++ mnist_recognition/src/model/dataset/mod.rs
@@ -1,21 +1,20 @@
-
 use mnist::*;
-use ndarray::{Array2, NdIndex, Dim};
+use ndarray::Array2;
 use rand::Rng;
 pub struct Dataset {
     // 2D array of flattened images from MNIST dataset, shuffled
-    pub training_data: Array2<f32>,
-    pub training_labels: Array2<f32>,
-    pub testing_data: Array2<f32>,
-    pub testing_labels: Array2<f32>,
+    pub training_data: Array2<f64>,
+    pub training_labels: Array2<f64>,
+    pub testing_data: Array2<f64>,
+    pub testing_labels: Array2<f64>,
 }
 
 impl Dataset {
-    fn shuffle(slices: &mut [&mut Array2<f32>]) {
-        if slices.len() > 0{
+    fn shuffle(slices: &mut [&mut Array2<f64>]) {
+        if slices.len() > 0 {
             let mut rng = rand::thread_rng();
             let shared_length = slices[0].index_axis_mut(ndarray::Axis(0), 0).len();
-            
+
             for i in 0..shared_length {
                 let next = rng.gen_range(i..shared_length);
 
@@ -25,7 +24,6 @@ impl Dataset {
                 }
             }
         }
-        
     }
 
     pub fn new() -> Dataset {
@@ -41,35 +39,33 @@ impl Dataset {
             .validation_set_length(10_000)
             .test_set_length(10_000)
             .finalize();
-        
-        
+
         // Create
-        let mut training_data: Array2<f32> = Array2::from_shape_vec((50_000, 784), trn_img)
+        let mut training_data: Array2<f64> = Array2::from_shape_vec((50_000, 784), trn_img)
             .expect("Error converting images to Array3 struct")
             .t()
-            .map(|x| *x as f32 / 256.0);
-    
+            .map(|x| *x as f64 / 256.0);
+
         // Convert the returned Mnist struct to Array2 format
-        let mut training_labels: Array2<f32> = Array2::from_shape_vec((50_000, 1), trn_lbl)
+        let mut training_labels: Array2<f64> = Array2::from_shape_vec((50_000, 1), trn_lbl)
             .expect("Error converting training labels to Array2 struct")
             .t()
-            .map(|x| *x as f32);
-    
-        let mut testing_data: Array2<f32> = Array2::from_shape_vec((10_000, 784), tst_img)
+            .map(|x| *x as f64);
+
+        let mut testing_data: Array2<f64> = Array2::from_shape_vec((10_000, 784), tst_img)
             .expect("Error converting images to Array3 struct")
             .t()
-            .map(|x| *x as f32 / 256.);
-    
-        let mut testing_labels: Array2<f32> = Array2::from_shape_vec((10_000, 1), tst_lbl)
+            .map(|x| *x as f64 / 256.);
+
+        let mut testing_labels: Array2<f64> = Array2::from_shape_vec((10_000, 1), tst_lbl)
             .expect("Error converting testing labels to Array2 struct")
             .t()
-            .map(|x| *x as f32);
-            
+            .map(|x| *x as f64);
+
         Dataset::shuffle(&mut [&mut training_data, &mut training_labels]);
         Dataset::shuffle(&mut [&mut testing_data, &mut testing_labels]);
-        
 
-        Dataset{ 
+        Dataset {
             training_data: training_data,
             training_labels: training_labels,
             testing_data: testing_data,
@@ -77,4 +73,3 @@ impl Dataset {
         }
     }
 }
-
